refactor(email): use Filament action for reply/forward buttons
Replace hand-rolled HTML buttons in email view with the shared
replyForwardEmailAction, registered as a modal action on ViewAction.
Make the action public and add link/icon styling per mode.

Also cast EMAIL_SYNC_INITIAL_DAYS and EMAIL_SYNC_BATCH_SIZE env values
to int.

+++ packages/EmailIntegration/config/email-integration.php
@@ -38,9 +38,9 @@ return [
      * Sync settings — override via .env
      */
     'sync' => [
-        'initial_days' => env('EMAIL_SYNC_INITIAL_DAYS', 90),
+        'initial_days' => (int) env('EMAIL_SYNC_INITIAL_DAYS', 90),
         'interval_minutes' => env('EMAIL_SYNC_INTERVAL_MINUTES', 5),
-        'batch_size' => env('EMAIL_SYNC_BATCH_SIZE', 50),
+        'batch_size' => (int) env('EMAIL_SYNC_BATCH_SIZE', 50),
     ],
 
     /*
